+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unrelease
 
+## Release
+
 ### 5.0.0
 
 - Update dev environment
@@ -13,8 +15,6 @@
 - Add redux store
 - Move actions and reducers from `twreporter-react` to this package
 - Remove unused reducers and actions
-
-## Release
 
 ### 4.0.3